Replace NumPy/stdlib names removed upstream
  np.float                 -> float            (removed in NumPy 1.24)
  np.in1d                  -> np.isin          (removed in NumPy 2.0)
  np.int                   -> int              (removed in NumPy 1.24)

These raise rather than warn, so they stop execution on a
current environment. The replacements are exact equivalents -
no behavioural or numerical change.

Touches 13 file(s); all still compile.

+++ python/fedml/simulation/mpi/fedgkt/utils.py
@@ -128,7 +128,7 @@ def save_dict_to_json(d, json_path):
     """Saves dict of floats in json file
 
     Args:
-        d: (dict) of float-castable values (np.float, int, float, etc.)
+        d: (dict) of float-castable values (float, int, float, etc.)
         json_path: (string) path to json file
     """
     with open(json_path, "w") as f:
